Refactor buckets.rb to file-per-method Add archived generations samples
