+++ google-cloud-storage/samples/storage_create_bucket.rb
@@ -1,0 +1,28 @@
+# Copyright 2020 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+def create_bucket bucket_name:
+  # [START storage_create_bucket]
+  # bucket_name = "Name of Google Cloud Storage bucket to create"
+
+  require "google/cloud/storage"
+
+  storage = Google::Cloud::Storage.new
+  bucket  = storage.create_bucket bucket_name
+
+  puts "Created bucket: #{bucket.name}"
+  # [END storage_create_bucket]
+end
+
+create_bucket bucket_name: ARGV.shift if $PROGRAM_NAME == __FILE__
