+++ google-cloud-storage/samples/storage_get_retention_policy.rb
@@ -1,0 +1,31 @@
+# Copyright 2020 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+def get_retention_policy bucket_name:
+  # [START storage_get_retention_policy]
+  # bucket_name = "Name of your Google Cloud Storage bucket"
+
+  require "google/cloud/storage"
+
+  storage = Google::Cloud::Storage.new
+  bucket  = storage.bucket bucket_name
+
+  puts "Retention policy:"
+  puts "period: #{bucket.retention_period}"
+  puts "effective time: #{bucket.retention_effective_at}"
+  puts "policy locked: #{bucket.retention_policy_locked?}"
+  # [END storage_get_retention_policy]
+end
+
+get_retention_policy bucket_name: ARGV.shift if $PROGRAM_NAME == __FILE__
