+++ google-cloud-storage/samples/storage_set_bucket_default_kms_key.rb
@@ -1,0 +1,31 @@
+# Copyright 2020 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+def set_bucket_default_kms_key bucket_name:, default_kms_key:
+  # [START storage_set_bucket_default_kms_key]
+  # bucket_name     = "Name of your Google Cloud Storage bucket"
+  # default_kms_key = "KMS key resource id"
+
+  require "google/cloud/storage"
+
+  storage = Google::Cloud::Storage.new
+  bucket  = storage.bucket bucket_name
+
+  bucket.default_kms_key = default_kms_key
+
+  puts "Default KMS key for #{bucket.name} was set to #{bucket.default_kms_key}"
+  # [END storage_set_bucket_default_kms_key]
+end
+
+set_bucket_default_kms_key bucket_name: ARGV.shift, default_kms_key: ARGV.shift if $PROGRAM_NAME == __FILE__
